(tgl)6/24/97: changes for HOF stars adjacent to galaxy boundaries.

+++ ptclassic/src/domains/hof/stars/HOFIfElseGr.pl
@@ -2,6 +2,15 @@ defstar {
 	name { IfElseGr }
 	domain { HOF }
 	derivedFrom { IfElse }
+	version { $Id$ }
+	author { Edward A. Lee, Tom Lane }
+	location { HOF main library }
+	copyright {
+Copyright (c) 1994-%Q% The Regents of the University of California.
+All rights reserved.
+See the file $PTOLEMY/copyright for copyright notice,
+limitation of liability, and disclaimer of warranty provisions.
+	}
 	desc {
 A variant of the IfElse star where the two possible replacement
 blocks are specified graphically rather than textually.
@@ -15,18 +24,9 @@ See the documentation for the
 <tt>IfElse</tt>
 star, from which this is derived, for background information.
 The parameter values for the replacement blocks are set directly
-by editing the parameters of the block.
+by editing the parameters of the blocks.
         }
-	version { $Id$ }
-	author { Edward A. Lee }
-	copyright {
-Copyright (c) 1990-%Q% The Regents of the University of California.
-All rights reserved.
-See the file $PTOLEMY/copyright for copyright notice,
-limitation of liability, and disclaimer of warranty provisions.
-	}
-	location { HOF main library }
-	ccinclude { "InfString.h" }
+
 	outmulti {
 	  name {trueout}
 	  type {anytype}
@@ -47,6 +47,9 @@ limitation of liability, and disclaimer of warranty provisions.
 	  type {anytype}
 	  desc {input from the false block}
 	}
+
+	ccinclude { "Galaxy.h" }
+
 	constructor {
 	  // The constructor for IfElse hides blockname, where_defined,
 	  // input_map, output_map, and parameter_map.  Here, we need to hide
@@ -62,26 +65,19 @@ limitation of liability, and disclaimer of warranty provisions.
 	  false_output_map.clearAttributes(A_SETTABLE);
 	  false_parameter_map.clearAttributes(A_SETTABLE);
 	}
-	protected {
-	  Block *trueblock, *falseblock;
-	}
+
 	method {
-	  name { preinitialize }
-	  access { public }
+	  name { doExpansion }
+	  type { int }
 	  code {
-	    HOFBaseHiOrdFn::preinitialize();
-
-	    Galaxy* mom = idParent();
-	    if(!mom) return;
-
 	    // Make sure we know the number of connections on the
 	    // input and output multiPortHoles.
-	    initConnections(input);
-	    initConnections(output);
-	    initConnections(trueout);
-	    initConnections(truein);
-	    initConnections(falseout);
-	    initConnections(falsein);
+	    if (! initConnections(input)) return 0;
+	    if (! initConnections(output)) return 0;
+	    if (! initConnections(truein)) return 0;
+	    if (! initConnections(trueout)) return 0;
+	    if (! initConnections(falsein)) return 0;
+	    if (! initConnections(falseout)) return 0;
 
 	    // At this point, any HOFNop stars will have been disconnected,
 	    // so there should be only one block connected to each of the
@@ -89,24 +85,32 @@ limitation of liability, and disclaimer of warranty provisions.
 
 	    MPHIter nexti(input);
 	    MPHIter nexto(output);
+	    MPHIter nexttruein(truein);
+	    MPHIter nexttrueout(trueout);
+	    MPHIter nextfalsein(falsein);
+	    MPHIter nextfalseout(falseout);
+
+	    // Choose the porthole over which to iterate
 	    MPHIter *nextexo;
 	    MPHIter *nextexi;
 	    MPHIter *nextnuo;
 	    MPHIter *nextnui;
-	    // Choose the porthole over which to iterate
 	    if (int(condition)) {
 	      // Iterators for ports connected to the block we will use
-	      LOG_NEW; nextexo = new MPHIter(trueout);
-	      LOG_NEW; nextexi = new MPHIter(truein);
+	      nextexo = &nexttrueout;
+	      nextexi = &nexttruein;
 	      // Iterators for ports connected to the block we won't use
-	      LOG_NEW; nextnuo = new MPHIter(falseout);
-	      LOG_NEW; nextnui = new MPHIter(falsein);
+	      nextnuo = &nextfalseout;
+	      nextnui = &nextfalsein;
 	    } else {
-	      LOG_NEW; nextexo = new MPHIter(falseout);
-	      LOG_NEW; nextexi = new MPHIter(falsein);
-	      LOG_NEW; nextnuo = new MPHIter(trueout);
-	      LOG_NEW; nextnui = new MPHIter(truein);
+	      nextexo = &nextfalseout;
+	      nextexi = &nextfalsein;
+	      nextnuo = &nexttrueout;
+	      nextnui = &nexttruein;
 	    }
+
+	    Galaxy* mom = idParent();
+	    if (!mom) return 0;
 
 	    PortHole *pi, *po, *pei, *peo;
 	    GenericPort *sourcegp, *destgp;
@@ -119,30 +123,43 @@ limitation of liability, and disclaimer of warranty provisions.
 	      if ((pi = nexti++) == 0) {
 		Error::abortRun(*this,
 				"Not enough inputs for the specified replacement block");
-		return;
+		return 0;
 	      }
-	      if (!(destgp = breakConnection(peo))) return;
-	      if (!connectInput(pi,destgp)) return;
+	      if (!(destgp = breakConnection(peo))) return 0;
+	      if (!connectInput(pi,destgp)) return 0;
 	    }
-	    // Better be out of outputs at this point
+	    // Better be out of inputs at this point
 	    if ((pi = nexti++) != 0) {
-	      Error::abortRun(*this, "Leftover inputs after connecting the replacement block");
-	      return;
+	      Error::abortRun(*this,
+			      "Leftover inputs after connecting the replacement block");
+	      return 0;
 	    }
 	    // Iterate over the outputs next
 	    while ((pei = (*nextexi)++) != 0) {
 	      if ((po = nexto++) == 0) {
 		Error::abortRun(*this,
 				"Not enough outputs for the specified replacement block");
-		return;
+		return 0;
 	      }
-	      if (!(sourcegp = breakConnection(pei))) return;
-	      if (!connectOutput(po, sourcegp)) return;
+	      if (!(sourcegp = breakConnection(pei))) return 0;
+	      if (!connectOutput(po, sourcegp)) return 0;
 	    }
+	    // Better be out of outputs at this point
+	    if ((po = nexto++) != 0) {
+	      Error::abortRun(*this,
+			      "Leftover outputs after connecting the replacement block");
+	      return 0;
+	    }
+
 	    if (!myblock) {
 	      Error::abortRun(*this,
 			      "No connections to replacement block!");
-	      return;
+	      return 0;
+	    }
+	    if (myblock->parent() != mom) {
+	      Error::abortRun(*this,
+			      "Replacement block is in wrong galaxy");
+	      return 0;
 	    }
 
 	    myblock = 0;	// now get a reference to the unused block
@@ -154,22 +171,23 @@ limitation of liability, and disclaimer of warranty provisions.
 	    while ((pei = (*nextnui)++) != 0) {
 	      breakConnection(pei);
 	    }
+
 	    if (!myblock) {
 	      Error::abortRun(*this,
 			      "No connections to unused replacement block!");
-	      return;
+	      return 0;
 	    }
-
-	    LOG_DEL; delete nextexo;
-	    LOG_DEL; delete nextexi;
-	    LOG_DEL; delete nextnuo;
-	    LOG_DEL; delete nextnui;
+	    if (myblock->parent() != mom) {
+	      Error::abortRun(*this,
+			      "Unused replacement block is in wrong galaxy");
+	      return 0;
+	    }
 
 	    // Delete the unused block from the parent galaxy.
 	    // This is *critical* to stop a graphical recursion.
 	    mom->deleteBlockAfterInit(*myblock);
 
-	    mom->deleteBlockAfterInit(*this);
+	    return 1;
 	  }
 	}
 }
